+++ src/NewDatePicker/index.ts
@@ -1,0 +1,2 @@
+export { default as NewDatePicker } from "./NewDatePicker";
+export type { NewDatePickerProps } from "./types";